+++ src/views/SimulatorView.vue
@@ -9,6 +9,7 @@
 import { storeToRefs } from "pinia";
 import { watch } from "vue";
 import { useTaxesStore } from "@/store";
+import { consumeInternalQueryUpdate } from "@/router";
 import { useRoute } from "vue-router";
 
 import Form from "@/components/Form.vue";
@@ -24,6 +25,10 @@ const route = useRoute();
 watch(
   () => route.fullPath,
   () => {
+    if (consumeInternalQueryUpdate()) {
+      return;
+    }
+
     store.setParametersFromURL(route.query);
   },
   { immediate: true },